+++ src/main/java/org/pcmm/gates/IGateID.java
@@ -11,8 +11,8 @@ import org.pcmm.base.IPCMMBaseObject;
  */
 public interface IGateID extends IPCMMBaseObject {
 	static final short LENGTH = 8;
-	static final short SNUM = 4;
-	static final short STYPE = 1;
+	static final byte SNUM = 4;
+	static final byte STYPE = 1;
 
 	void setGateID(int gateID);
