+++ src/main/java/org/pcmm/gates/IIPv6Classifier.java
@@ -2,8 +2,8 @@ package org.pcmm.gates;
 
 public interface IIPv6Classifier extends IExtendedClassifier {
 	static final short LENGTH = 64;
-	static final short SNUM = 6;
-	static final short STYPE = 3;
+	static final byte SNUM = 6;
+	static final byte STYPE = 3;
 
 	// Tc-low
 	// Tc-high
