+++ src/main/java/org/pcmm/gates/ITrafficProfile.java
@@ -3,7 +3,7 @@ package org.pcmm.gates;
 import org.pcmm.base.IPCMMBaseObject;
 
 public interface ITrafficProfile extends IPCMMBaseObject {
-	static final short SNUM = 7;
+	static final byte SNUM = 7;
 
 	/**
 	 * 0x001, 0x011 and 0x111 (Authorized, Reserved, and Committed) are allowed
